Split presentation cache lifecycle
