+++ CMakeLists.txt
@@ -210,6 +210,7 @@ target_sources(
         src/game_data_standard_options.c
         src/game_presentation.c
         src/game_presentation_app_flow.c
+        src/game_presentation_cache_lifecycle.c
         src/game_presentation_menus.c
         src/game_presentation_scene_flow.c
         src/game_data_validation_actions.c
